+++ ServersideQoL.ContainerSigns/CHANGELOG.md
@@ -1,2 +1,5 @@
+### v2.0.4
+- Fixed signs do not appear when the container size is also changed [#190](https://github.com/ArgusMagnus/ValheimServersideQoL/issues/190)
+
 ### v2.0.0
 - Initial release