+++ ServersideQoL.CreatureLevelUp/CHANGELOG.md
@@ -1,2 +1,5 @@
+### v2.0.4
+- Minor fix related to [#190](https://github.com/ArgusMagnus/ValheimServersideQoL/issues/190)
+
 ### v2.0.0
 - Initial release